+++ base/gxclist.c
@@ -353,7 +353,7 @@ clist_init_states(gx_device * dev, byte * init_data, uint data_size)
         &((gx_device_clist *)dev)->writer;
     ulong state_size = cdev->nbands * (ulong) sizeof(gx_clist_state);
     /* Align to the natural boundary for ARM processors, bug 689600 */
-    long alignment = (-(long)init_data) & (sizeof(init_data) - 1);
+    intptr_t alignment = (-(intptr_t)init_data) & (sizeof(init_data) - 1);
 
     /* Leave enough room after states for commands that write a reasonable
      * amount of data. The cmd_largest_size and the data_bits_size should  be
